% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{annot}
\alias{annot}
\title{Annotate variant data frame.}
\usage{
annot(x, y, type, nomatch, CLC)
}
\arguments{
\item{x}{the data frame to be annotated.}

\item{y}{the data frame containing the annotation.}

\item{type}{specify whether y is a BED or GTF file.}

\item{nomatch}{should variants located outside any of the regions found in the
annotation data frame be removed (FALSE) or withheld (TRUE).}

\item{CLC}{logical, origin of the data frame to be annotated is CLC genomics
workbench?}
}
\description{
annot can be used to annotate a data frame containing sequencing
 variants.
}
\details{
This function can be used to annotate data frames. For further
 processing by gDom, gRec and commonvar, annotation is necessary. For nDom
 and nRec, annotation is optional. Both variant files from CLC Genomics
 Workbench and VCF files (after VCFfile) can be processed. If a variant can
 be allocated to >1 group (a gene, exon or other), the variant information is
 spread over n rows (with n the number of groups).
}
\examples{
data(CLCfile1)
 data(genBED)
 # produces an annotated CLC file with all variants retained.
 AnnotCLCfile1 <- annot(CLCfile1, genBED, type="BED", nomatch=TRUE, CLC=TRUE)
 # produces an annotated CLC file with only the variants that were allocated
 # to a gene being retained.
 AnnotCLCfile2 <- annot(CLCfile1, genBED, type="BED", nomatch=FALSE, CLC=TRUE)
}
\author{
Bart Broeckx
}

