% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CLCfile}
\alias{CLCfile}
\title{Prepare CLC output for filtering}
\usage{
CLCfile(x, y, multiple, split)
}
\arguments{
\item{x}{the data frame to be processed.}

\item{y}{the column name containing the grouping variable.}

\item{multiple}{does one line contain >1 group? If true, specify split.}

\item{split}{the split that separates the group in one line.}
}
\description{
CLCfile should be used to process CLC output prior downstream
  analysis with gDom and gRec.
}
\details{
This function is specifically designed to process the output from
  CLC genomics workbench. It should be used before the variant filters gDom
  and gRec are used to emphasize the grouping (e.g. gene, exon) to be used in
  downstream analysis. As often is the case, a variant might belong to >1
  group (e.g. 2 genes). If so, the file is automatically reformatted to
  assure that each line contains only one variant per group, if multiple is
  TRUE and the split is specified. If multiple is FALSE, only the first part
  of the y-column is maintained.This function should not be used for nDom and
  nRec.
}
\examples{
CLCfile1proc <-CLCfile(CLCfile1, "Coding.region.change", TRUE, "; ")
CLCfile2proc <-CLCfile(CLCfile2, "Coding.region.change", TRUE, "; ")
}
\author{
Bart Broeckx
}

