% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CLCfile1}
\alias{CLCfile1}
\title{CLCfile1 documentation
A dataset containing 8 variants (output CLC genomics workbench)}
\format{\preformatted{'data.frame':	8 obs. of  21 variables:
 $ Chromosome                 : chr  "chr1" "chr1" "chr1" "chr1" ...
 $ Region                     : chr  "266523" "363618" "604894" "699873" ...
 $ Type                       : chr  "SNV" "SNV" "SNV" "SNV" ...
 $ Reference                  : chr  "G" "A" "A" "G" ...
 $ Allele                     : chr  "T" "G" "G" "A" ...
 $ Reference.allele           : chr  "No" "No" "No" "No" ...
 $ Length                     : int  1 1 1 1 1 1 1 3
 $ Linkage                    : logi  NA NA NA NA NA NA ...
 $ Zygosity                   : chr  "Heterozygous" "Heterozygous" "Homozygous" "Homozygous" ...
 $ Count                      : int  12 27 212 21 23 15 22 5
 $ Coverage                   : int  24 52 212 21 23 61 62 5
 $ Frequency                  : num  50 51.9 100 100 100 ...
 $ Probability                : num  1 1 1 1 1 1 1 1
 $ Forward.read.count         : int  8 27 134 17 23 7 5 1
 $ Reverse.read.count         : int  5 3 87 7 2 9 17 4
 $ Forward.reverse.balance    : num  0.385 0.1 0.394 0.292 0.08 0.438 0.227 0.2
 $ Average.quality            : num  38 36.1 36.2 35.1 33 ...
 $ Coding.region.change       : chr  "ENSCAFT00000000001:c.1802C>A" "ENSCAFT00000000003:c.12A>G" "ENSCAFT00000000008:c.775T>C" "ENSCAFT00000037436:c.371C>T" ...
 $ Amino.acid.change          : chr  "ENSCAFT00000000001:p.Pro601His" "ENSCAFT00000000003:p.Ile4Met" "ENSCAFT00000000008:p.Phe259Leu" "ENSCAFT00000037436:p.Ala124Val" ...
 $ Other.variants.within.codon: chr  "No" "No" "No" "No" ...
 $ Non.synonymous             : chr  "Yes" "Yes" "Yes" "Yes" ...
}}
\usage{
CLCfile1
}
\description{
CLCfile1 documentation
A dataset containing 8 variants (output CLC genomics workbench)
}
\keyword{datasets}

