% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{CLCfile2}
\alias{CLCfile2}
\title{CLCfile2 documentation
A dataset containing 9 variants (output CLC genomics workbench)}
\format{\preformatted{'data.frame':	9 obs. of  21 variables:
 $ Chromosome                 : chr  "chr1" "chr1" "chr1" "chr1" ...
 $ Region                     : chr  "266523" "604894" "699873" "758461^758462" ...
 $ Type                       : chr  "SNV" "SNV" "SNV" "Insertion" ...
 $ Reference                  : chr  "G" "A" "G" "-" ...
 $ Allele                     : chr  "T" "G" "A" "G" ...
 $ Reference.allele           : chr  "No" "No" "No" "No" ...
 $ Length                     : int  1 1 1 1 1 1 1 1 1
 $ Linkage                    : logi  NA NA NA NA NA NA ...
 $ Zygosity                   : chr  "Heterozygous" "Homozygous" "Homozygous" "Homozygous" ...
 $ Count                      : int  7 151 13 23 23 18 30 124 37
 $ Coverage                   : int  16 151 16 23 23 67 69 124 37
 $ Frequency                  : num  43.8 100 81.2 100 100 ...
 $ Probability                : num  1 1 0.78 1 1 1 1 1 1
 $ Forward.read.count         : int  6 86 11 23 23 10 8 28 32
 $ Reverse.read.count         : int  2 73 3 2 2 12 22 103 11
 $ Forward.reverse.balance    : num  0.25 0.459 0.214 0.08 0.08 0.455 0.267 0.214 0.256
 $ Average.quality            : num  34 36 35.9 37.3 38 ...
 $ Coding.region.change       : chr  "ENSCAFT00000000001:c.1802C>A" "ENSCAFT00000000008:c.775T>C" "ENSCAFT00000037436:c.371C>T" "ENSCAFT00000000011:c.48_49insG" ...
 $ Amino.acid.change          : chr  "ENSCAFT00000000001:p.Pro601His" "ENSCAFT00000000008:p.Phe259Leu" "ENSCAFT00000037436:p.Ala124Val" "ENSCAFT00000000011:p.Val17fs" ...
 $ Other.variants.within.codon: chr  "No" "No" "No" "No" ...
 $ Non.synonymous             : chr  "Yes" "Yes" "Yes" "Yes" ...
}}
\usage{
CLCfile2
}
\description{
CLCfile2 documentation
A dataset containing 9 variants (output CLC genomics workbench)
}
\keyword{datasets}

