% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{commonvar}
\alias{commonvar}
\title{Combining the output of several filters.}
\usage{
commonvar(x, group)
}
\arguments{
\item{x}{the data frames to be filtered (> 1).}

\item{group}{the column names of the columns containing the functional units.}
}
\description{
commonvar should be used to combine the output of nDom/nRec and
 gDom/gRec filters.
}
\details{
This function allows the combination of nDom/nRec and gDom/gRec
 output. It retains all functional units in common, whithout assumptions on
 zygosity (as this should already have been taken care of in the previous
 filtering steps). Within each functional unit, the variants present in each
 x object are retained. Detectance can be specified and will be used on the
 functional group instead of the individual variant.
}
\examples{
output <- nRec("CLCfile1", "CLCfile2"  )
output
CLCfile1proc <-CLCfile(CLCfile1, "Coding.region.change", TRUE, "; ")
CLCfile2proc <-CLCfile(CLCfile2, "Coding.region.change", TRUE, "; ")
output2 <- gDom("CLCfile1proc", "CLCfile2proc")
output2
Annotoutput2 <- annot(output,genBED, type="BED", nomatch=FALSE, CLC=TRUE)
x <- c("output2", "Annotoutput2")
group <- c("Group", "annotation")
a <-commonvar(x,group)
}
\author{
Bart Broeckx
}

