% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{gDom}
\alias{gDom}
\title{Variant filtering with dominant inheritance considering functional unit.}
\usage{
gDom(x, y)
}
\arguments{
\item{x}{the case(s).}

\item{y}{optional, the control(s).}
}
\description{
gDom should be used to filter variants under a dominant mode of
  inheritance when common functional units (a gene, an exon, ...) should be considered.
}
\details{
gDom checks for common functional units within cases. If variants from
  control(s) are present, they can be used to filter variants from the cases.
  In addition, various detectance and penetrance parameters can be specified
  during the processing.

  Before using this function, each file from each case should have been processed by CLCfile or VCFfile with annot.
}
\examples{
CLCfile1proc <-CLCfile(CLCfile1, "Coding.region.change", TRUE, "; ")
CLCfile2proc <-CLCfile(CLCfile2, "Coding.region.change", TRUE, "; ")
 output <- gDom("CLCfile1proc", "CLCfile2proc")
 output
 output <- gDom(c("CLCfile1proc","CLCfile2proc"))
 output
}
\author{
Bart Broeckx
}

