% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{genBED}
\alias{genBED}
\title{genBED documentation
A bed file containing genes used to annotate files}
\format{\preformatted{'data.frame':	1813 obs. of  12 variables:
 $ V1 : Factor w/ 1 level "chr1": 1 1 1 1 1 1 1 1 1 1 ...
 $ V2 : int  247828 270004 363606 509124 571408 600161 692601 722178 744460 757902 ...
 $ V3 : int  322180 275409 364548 565905 571749 635340 722173 735934 746178 800885 ...
 $ V4 : Factor w/ 1813 levels "ENSCAFT00000000001",..: 1 1382 2 3 4 5 1152 6 1469 7 ...
 $ V5 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V6 : Factor w/ 2 levels "-","+": 1 2 2 2 2 1 1 2 1 2 ...
 $ V7 : int  252391 270004 363606 509124 571749 602263 693293 722724 744460 757902 ...
 $ V8 : int  322063 275367 364548 565905 571749 625146 703694 735009 746178 800885 ...
 $ V9 : int  0 0 0 0 0 0 0 0 0 0 ...
 $ V10: int  25 2 1 4 2 5 7 4 2 7 ...
 $ V11: Factor w/ 1680 levels "10,116,69,162,132,88,138,122,157,147,353,",..: 1148 56 1627 325 758 1227 41 303 1608 1079 ...
 $ V12: Factor w/ 1466 levels "0,","0,100,",..: 1205 948 1 505 634 1438 517 1050 329 1096 ...
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables?command=start}
}
\usage{
genBED
}
\description{
genBED documentation
A bed file containing genes used to annotate files
}
\keyword{datasets}

