% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gRec}
\alias{gRec}
\title{Variant filtering with recessive inheritance considering functional unit.}
\usage{
gRec(x, y)
}
\arguments{
\item{x}{the case(s).}

\item{y}{optional, the control(s).}
}
\description{
gRec should be used to filter variants under a recessive mode of
 inheritance when common functional units (a gene, an exon, ...) should be
 considered.
}
\details{
gRec checks for common functional units within cases. For a unit to
 be retained, it should have at least one homozygous non-reference variant or
 at least be compound heterozygous.  If variants from control(s) are present,
 they can be used to filter variants from the cases. Homozygous variants
 present in the controls will be used to filter variants from the cases
 immediately. Pairwise combinations of heterozygous variants present in the
 controls can also be used to filter variants in the cases. In addition,
 various detectance and penetrance parameters can be specified during the
 processing.

 Before using this function, each file from each case should have been
 processed by CLCfile or VCFfile with annot.
}
\examples{
CLCfile1proc <-CLCfile(CLCfile1, "Coding.region.change", TRUE, "; ")
CLCfile2proc <-CLCfile(CLCfile2, "Coding.region.change", TRUE, "; ")
output <- gRec("CLCfile1proc", "CLCfile2proc")
output
output <- gRec(c("CLCfile1proc","CLCfile2proc"))
output
}
\author{
Bart Broeckx
}

