% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nDom}
\alias{nDom}
\title{Variant filtering with dominant inheritance}
\usage{
nDom(x, y, family)
}
\arguments{
\item{x}{the case(s).}

\item{y}{optional, the control(s).}

\item{family}{optional, "P-F" or "Ps-F" mode, indicating a (healthy)
parent(s)- (affected) progeny relation.}
}
\description{
nDom should be used to filter variants under a dominant mode of
  inheritance.
}
\details{
nDom checks for common variants within cases. If variants from
  control(s) are present, they can be used to filter variants from the cases.
  In addition, various detectance and penetrance parameters can be specified
  during the processing.

  If family is specified, penetrance is assumed to be complete. In addition,
  the causal variant is assumed to be heterozygous in the case.
}
\examples{
output <- nDom("CLCfile1", "CLCfile2"  )
output
output <- nDom(c("CLCfile1","CLCfile2"))
output
}
\author{
Bart Broeckx
}

