% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nRec}
\alias{nRec}
\title{Variant filtering with recessive inheritance}
\usage{
nRec(x, y, family)
}
\arguments{
\item{x}{the case(s).}

\item{y}{optional, the control(s).}

\item{family}{optional, "P-F" or "Ps-F" mode, indicating a (healthy)
parent(s)- (affected) progeny relation.}
}
\description{
nRec should be used to filter variants under a recessive mode of
  inheritance.
}
\details{
nRec checks for common variants that are homozygous in the cases. If variants from
  control(s) are present, they can be used to filter variants from the cases. However, only homozygous variants are used to filter.
  In addition, various detectance and penetrance parameters can be specified
  during the processing.

  If family is specified, penetrance is assumed to be complete. In addition,
  the causal variant is assumed to be homozygous in the case and heterozygous in the parent(s).
}
\examples{
output <- nRec("CLCfile1", "CLCfile2")
output
output <- nRec(c("CLCfile1","CLCfile2"))
output
}
\author{
Bart Broeckx
}

