% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{prepvar}
\alias{prepvar}
\title{Preparing variant database.}
\usage{
prepvar(y, MAF, reference)
}
\arguments{
\item{y}{the data frame containing the variants.}

\item{MAF}{optional numeric, the MAF of the non-reference variants to be
retained for filtering.}

\item{reference}{the column containing the reference sequence in the y data
frame.}
}
\description{
prepvar is used to prepare a variant database for further
 filtering with varfilter.
}
\details{
This function prepares the variants from a database like dbSNP to be
 used for filtering variants in cases. MAF is optional and specifies a
 minimum allele frequency for the database variants to be retained for
 variant filtering. The input of y should be the standard output of the UCSC
 table browser ("all fields from selected table"). For large files, running
 time can be quite long. An alternative supporting parallel computing is
 prepvarpar. For the actual filtering, use varfilter.
}
\examples{
y <- SNP
reference <- "refNCBI"
MAF <- 0.03
a <-prepvar(y, MAF, reference)
}
\author{
Bart Broeckx
}

