% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{prepvarpar}
\alias{prepvarpar}
\title{Preparing variant database (parallel).}
\usage{
prepvarpar(y, MAF, reference, nproc)
}
\arguments{
\item{y}{the data frame containing the variants.}

\item{MAF}{optional numeric, the MAF of the non-reference variants to be
retained for filtering.}

\item{reference}{the column containing the reference sequence in the y data
frame.}

\item{nproc}{the number of cores or workers used for parallel computing.}
}
\description{
prepvarpar is used to prepare a large variant database for
 further filtering with varfilter.
}
\details{
This function prepares the variants from a large database like dbSNP
 to be used for filtering variants in cases. MAF is optional and specifies a
 minimum allele frequency for the database variants to be retained for
 variant filtering. The input of y should be the standard output of the UCSC
 table browser ("all fields from selected table"). This function allows
 parallel computing, leading to substantial time improvement when entire
 databases are used. If parallel computing is not necessary or not possible,
 use prepvar.Prior to running this function, call function registerDoParallel
 from the doParallel package to permit parallel computing.
}
\examples{
#library(doParallel)
#registerDoParallel()
# nproc <-getDoParWorkers()
# a <-prepvarpar(SNP, 0.03,"refNCBI", 1)
}

