% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{SNP}
\alias{SNP}
\title{SNP documentation
A dataset containing 472 variants from the dbSNP database}
\format{\preformatted{'data.frame':	472 obs. of  26 variables:
 $ bin            : int  596 596 596 596 596 596 596 596 596 596 ...
 $ chrom          : Factor w/ 1 level "chr1": 1 1 1 1 1 1 1 1 1 1 ...
 $ chromStart     : int  1500485 1500940 1501063 1501088 1501181 1501465 1501631 1501655 1501796 1501995 ...
 $ chromEnd       : int  1500502 1500941 1501064 1501089 1501182 1501466 1501632 1501655 1501803 1501996 ...
 $ name           : Factor w/ 472 levels "rs10157140","rs1061892",..: 117 331 337 420 438 406 135 214 114 116 ...
 $ score          : int  0 0 0 0 0 0 0 0 0 0 ...
 $ strand         : Factor w/ 2 levels "-","+": 2 2 2 2 2 2 2 2 2 2 ...
 $ refNCBI        : Factor w/ 22 levels "-","A","AAAG",..: 5 2 20 17 20 17 17 1 14 17 ...
 $ refUCSC        : Factor w/ 22 levels "-","A","AAAG",..: 5 2 20 17 20 17 17 1 14 17 ...
 $ observed       : Factor w/ 36 levels "-/A","-/A/T",..: 5 32 36 32 35 32 32 1 14 32 ...
 $ molType        : Factor w/ 2 levels "cDNA","genomic": 2 2 2 2 2 2 2 2 2 2 ...
 $ class          : Factor w/ 3 levels "deletion","insertion",..: 1 3 3 3 3 3 3 2 1 3 ...
 $ valid          : Factor w/ 13 levels "by-1000genomes",..: 3 9 9 6 6 9 1 3 3 3 ...
 $ avHet          : num  0.4777 0.4643 0.4012 0.0492 0.1377 ...
 $ avHetSE        : num  0.103 0.129 0.199 0.149 0.223 ...
 $ func           : Factor w/ 16 levels "coding-synon,missense",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ locType        : Factor w/ 3 levels "between","exact",..: 3 2 2 2 2 2 2 1 3 2 ...
 $ weight         : int  1 1 1 1 1 1 1 1 1 1 ...
 $ exceptions     : Factor w/ 8 levels "","DuplicateObserved",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ submitterCount : int  5 20 12 3 4 15 1 3 2 2 ...
 $ submitters     : Factor w/ 271 levels "1000GENOMES,",..: 163 2 131 221 216 31 1 240 238 254 ...
 $ alleleFreqCount: int  2 2 2 2 2 2 2 2 2 2 ...
 $ alleles        : Factor w/ 36 levels "-,A,","-,A,T,",..: 5 32 36 32 35 32 32 1 14 32 ...
 $ alleleNs       : Factor w/ 410 levels "1.000000,11.000000,",..: 43 86 290 367 132 60 215 194 270 311 ...
 $ alleleFreqs    : Factor w/ 410 levels "0.000457,0.028349,0.971193,",..: 215 153 252 50 79 228 5 373 15 35 ...
 $ bitfields      : Factor w/ 8 levels "","maf-5-some-pop",..: 1 4 4 4 2 4 1 1 1 1 ...
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables?command=start}
}
\usage{
SNP
}
\description{
SNP documentation
A dataset containing 472 variants from the dbSNP database
}
\keyword{datasets}

