% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{test}
\alias{test}
\title{test documentation
An example of a VCF file.}
\format{\preformatted{'data.frame':	21 obs. of  10 variables:
 $ V1 : Factor w/ 24 levels "chr1","chr10",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ V2 : int  1563789 1564952 1570039 1575616 1577003 1580738 1581065 1581096 1581558 1581586 ...
 $ V3 : Factor w/ 54924 levels ".","rs10000250",..: 47593 4922 1 51334 1 1 1 1 8702 40635 ...
 $ V4 : Factor w/ 693 levels "A","AAAAAAAAAAAAAAAAAAG",..: 176 609 367 527 176 527 367 176 176 527 ...
 $ V5 : Factor w/ 634 levels "A","A,AC","A,ACTGCTGCTG",..: 1 468 152 152 468 1 152 468 468 152 ...
 $ V6 : num  946 529 101 1530 119 ...
 $ V7 : Factor w/ 80 levels "HARD_TO_VALIDATE",..: 69 80 31 30 31 80 79 79 31 31 ...
 $ V8 : Factor w/ 61487 levels "AB=0.004;AC=1;AF=0.50;AN=2;BaseQRankSum=1.605;DB;DP=480;Dels=0.00;FS=8.451;HRun=0;HaplotypeScore=2.6000;MQ=4.51;MQ0=449;MQRankS"| __truncated__,..: 41114 45274 60986 58527 61394 28728 17694 16338 54678 53041 ...
 $ V9 : Factor w/ 1 level "GT:AD:DP:GQ:PL": 1 1 1 1 1 1 1 1 1 1 ...
 $ V10: Factor w/ 60111 levels "0/1:1,10:11:6.90:348,0,7",..: 55487 55182 54866 52753 58960 27959 17315 22508 49617 48313 ...
}}
\usage{
test
}
\description{
test documentation
An example of a VCF file.
}
\keyword{datasets}

