% Generated by roxygen2 (4.0.2): do not edit by hand
\name{varfilter}
\alias{varfilter}
\title{variant databse filtering from cases.}
\usage{
varfilter(x, y, MAF, reference)
}
\arguments{
\item{x}{the data frame to be filtered.}

\item{y}{the data frame containing the variants.}

\item{MAF}{optional numeric, the MAF of the non-reference variants to be
retained for filtering.}

\item{reference}{the column containing the reference sequence in the y data frame.}
}
\description{
varfilter can be used to filter case variants using a variant
 database (like dbSNP).
}
\details{
This function allows variant filtering from databases like dbSNP. MAF
 is optional and specifies a minimum allele frequency for the database variants to be
 retained for variant filtering in the object of interest (= x). The input of y
 should be the standard output of the UCSC table browser ("all fields from
 selected table").
}
\examples{
out <-VCFfile(test, "V10", TRUE, "PASS")
x <- out
y <- SNP
reference <- "refNCBI"
MAF <- 0.03
a <-varfilter(x,y, MAF, reference)
}
\author{
Bart Broeckx
}

