% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{varfilter}
\alias{varfilter}
\title{Filtering variants from a database.}
\usage{
varfilter(x, y)
}
\arguments{
\item{x}{the data frame to be filtered}

\item{y}{the data frame containing the variants.}
}
\description{
varfilter is used to filter case variants using a variant
 database (like dbSNP).
}
\details{
After preprocessing the variant database with prepvar or
 prepvarpar, the actual filtering is done with varfilter.
}
\examples{
out <-VCFfile(test, "V10", TRUE, "PASS")
a <-prepvar(SNP, 0.03,"refNCBI")
filtered <-varfilter(out,a)
}
\author{
Bart Broeckx
}

